feat(nether_wait): Use Nether Portals in survival without delay
Teleporting with Nether Portals in creative is instant, but a significant delay is incurred in survival. Internally, the portal wait time is tied to the player's invulnerability property. InstantNetherPortals, an earlier plugin for the same purpose removes the delay in survival by making the player invulnerable for an instant. This plugin uses the same workaround.

For the workaround to work, invulnerability must be given through the net.minecraft.server/CraftBukkit API. The downside is that CraftBukkit is a versioned package accessible through `org.bukkit.craftbukkit.<VERSION>`, which must be hardcoded (instead of being dynamically resolved based on the build environment). This may be worked around at some point later.

In addition to the functionality of the original plugin, teleportation happens after the normal delay when crouching in the Nether Portal in survival.